requirements.txt
dspy
agno
openpyxl
scikit-learn
streamlit-theme
pandas
numpy
plotly

